[ALPS03593880] scp: fix 32bit kernel build error
1. fix cast from pointer to integer of different size(ioremap_wc)
2. fix (struct log_ctrl_s *) point to 64bit address

MTK-Commit-Id: 121268ca7e556559b7997199db1122ef4fac4973

CR-Id: ALPS03593880
Feature: TinySys

+++ drivers/misc/mediatek/scp/v01/scp_helper.c
@@ -915,7 +915,7 @@ static int scp_reserve_memory_ioremap(void)
 		return -1;
 	}
 	accumlate_memory_size = 0;
-	scp_mem_base_virt = (phys_addr_t)ioremap_wc(scp_mem_base_phys, scp_mem_size);
+	scp_mem_base_virt = (phys_addr_t)(size_t)ioremap_wc(scp_mem_base_phys, scp_mem_size);
 	pr_debug("[SCP]reserve mem: virt:0x%llx - 0x%llx (0x%llx)\n", (phys_addr_t)scp_mem_base_virt,
 		(phys_addr_t)scp_mem_base_virt + (phys_addr_t)scp_mem_size, scp_mem_size);
 	for (id = 0; id < NUMS_MEM_ID; id++) {
